requirements.txt
dnspython>=2.7.0,<3.0.0
paramiko